addition Kletskov2019.pdf as a test file
Provision of a test file to probe the efficiency of the scripts'
action as an invite for replication.  Kletskov, A. V., Bumagin,
N. A., Zubkov, F. I.; Grudinin, D. G.; Potkin, V. I. 'Isothiazoles
in the Design and Synthesis of Biologically Active Substances and
Ligands for Metal Complexes', in Synthesis, 2020, 52, 159--188,
doi 10.1055/s-0039-1690688.  This is a review article published as
open access publication (license terms creative commons) by Thieme
Verlag Stuttgart, Germany.  Retrieved on 2019-12-17 (YYYY-MM-DD) on

https://www.thieme-connect.com/products/ejournals/abstract/10.1055/s-0039-1690688

as an example publication.