add Kletskov2019.pdf as test file
Provision of a reference file to check and demonstrate the
efficiency of the script.  Kletskov, A. V. et al. 'Isothiazoles
in the Design and Synthesis of Biologically Active Substances
and Ligands for Metal Complexes' in Synthesis, 2020, 52, 159-188,
doi 10.1055/s-0039-1690688 may serve as an example as it contains
text, line art, and photographs.  It was retrieved today on

https://www.thieme-connect.com/products/ejournals/abstract/10.1055/s-0039-1690688

as open access publication (no pay-wall) and the permissive terms
of the creative commons license.