interchange program parameter / file parameter
In comparison with other bash programs, it seems preferable to
provide the instruction what to do (reprint in color, or in
gray scale) prior to the instruction about which .pdf file to
process.  Equally, the script now accepts either short or long
toggles as in -c, --color, --reprint; or -g, --gray).