interchange program parameter / file parameter
In comparison with other bash programs, it seems preferable to
provide the instruction what to do (reprint in color, or in gray
scale) prior to the instruction about which .pdf to work on.  By
this commit, this interchange is installed.

Equally, the script now accepts either short or long toggles
(-c, --color, --reprint; or -g, --gray) for either work ahead.