bug fix (gray scale); add minimal -h menu
The conversion to gray scale is established again (syntax error).
A basic help menu is accessible when running ./pdf_rewrite.sh -h,
the user is hinted to this when running ./pdf_rewrite.sh alone.

Prooven to work both for color reduction and link retention with
Linux Debian 11 bullseye (branch testing) with ghostscript (9.53.3)
as processor and qpdfview (0.4.18) as viewer.